refactor: Replace old plugin implementations with refactored versions
- Replace Homebrew brewfile.rs with mixin-based implementation
- Replace NPM config.rs and global_packages.rs with mixin-based implementations
- Remove _new suffixes from type aliases and registration names
- Clean up module references and remove old implementation files
- Fix dead code warning in ValidationHelpers
- Maintain 117 tests passing (removed 24 redundant old tests)

Final result: Clean, consistent mixin-based architecture across all plugins
