+++ src/plugins/homebrew/mod.rs
@@ -1,5 +1,4 @@
 pub mod brewfile;
-pub mod brewfile_new;
 
 // Auto-registration means we don't need to export plugin types anymore
 // The inventory system handles plugin discovery automatically
