+++ src/plugins/npm/mod.rs
@@ -1,7 +1,5 @@
 pub mod config;
-pub mod config_new;
 pub mod global_packages;
-pub mod global_packages_new;
 
 // Auto-registration means we don't need to export plugin types anymore
 // The inventory system handles plugin discovery automatically
